altered BSP_ADC.c to user int32 values while also using mV instead of V to avoid precision loss. Clark then altered files which used the old Update_measurements function from float to int variable types. Finally, started two test files for ADC and CAN

+++ BSP/Inc/BSP_CAN.h
@@ -18,7 +18,7 @@ void BSP_CAN_Init(void);
  * @param   Voltage : the voltage value at that moment in a float
  * @return  0 if module was unable to transmit the data onto the CAN bus. Any other value indicates data was transmitted.
  */
-uint8_t BSP_CAN_Write(uint32_t id, float Voltage);
+uint8_t BSP_CAN_Write(uint32_t id, int32_t Voltage);
 
 /**
  * @brief   Gets the data that was received from the CAN bus.
